Offline Img for all

+++ Girl Frontline Info/Controller/OfflineSettingsViewController.swift
@@ -8,7 +8,7 @@
 
 import UIKit
 
-class OfflineSettingsViewController: UITableViewController, VersionProtocol, localDBDelegate, localDataDelegate {
+class OfflineSettingsViewController: UITableViewController, VersionProtocol, localDBDelegate, localDataDelegate, URLSessionDelegate {
 
     @IBOutlet weak var versionLabel: UILabel!
     @IBOutlet weak var tsLabel: UILabel!
@@ -28,7 +28,7 @@ class OfflineSettingsViewController: UITableViewController, VersionProtocol, loc
     let noti = UINotificationFeedbackGenerator()
     let tap = UISelectionFeedbackGenerator()
     let localPath = FileManager.default.urls(for: .documentDirectory, in: .userDomainMask)
-    let imgAlert = UIAlertController(title: "下載圖片", message: "這需要一分鐘至數分鐘時間，建議使用Wi-Fi網絡下載以避免流動數據收費。", preferredStyle: .alert)
+    let imgAlert = UIAlertController(title: "下載圖片", message: "這可能需要一分鐘至數分鐘時間，請耐心等候", preferredStyle: .alert)
     let delImgAlert = UIAlertController(title: "將刪除已下載圖片", message: nil, preferredStyle: .alert)
     
     var userDefaults = UserDefaults.standard
@@ -49,7 +49,6 @@ class OfflineSettingsViewController: UITableViewController, VersionProtocol, loc
                 alert.addAction(UIAlertAction(title: "確定", style: .default, handler: nil))
                 self.present(alert, animated: true, completion: nil)
                 noti.notificationOccurred(.warning)
-                downImgSwitch.isEnabled = true
             }else{
                 startAlert.view.addSubview(prog)
                 self.present(startAlert, animated: true, completion: nil)
@@ -70,7 +69,6 @@ class OfflineSettingsViewController: UITableViewController, VersionProtocol, loc
             self.present(alert, animated: true, completion: nil)
         })
         noti.notificationOccurred(.success)
-        downImgSwitch.isEnabled = true
         self.localSearch.readVersion()
     }
     func failed() {
@@ -130,7 +128,7 @@ class OfflineSettingsViewController: UITableViewController, VersionProtocol, loc
             if let mydb = db{
                 let statement = mydb.fetch("info", cond: nil, order: nil)
                 if sqlite3_step(statement) == SQLITE_ROW{
-                    let alert = UIAlertController(title: "更新資料檔", message: nil, preferredStyle: .alert)
+                    let alert = UIAlertController(title: "更新資料檔", message: "這可能需要一分鐘至數分鐘時間，請耐心等候", preferredStyle: .alert)
                     alert.addAction(UIAlertAction(title: "確定", style: .default, handler: { _ in
                         localDB().writeSettings(item: "isOffline", value: "1")
                         self.type = "update"
@@ -142,7 +140,7 @@ class OfflineSettingsViewController: UITableViewController, VersionProtocol, loc
                     self.present(alert, animated: true, completion: nil)
                     
                 }else{
-                    let alert = UIAlertController(title: "下載資料檔", message: nil, preferredStyle: .alert)
+                    let alert = UIAlertController(title: "下載資料檔", message: "這可能需要一分鐘至數分鐘時間，請耐心等候", preferredStyle: .alert)
                     alert.addAction(UIAlertAction(title: "確定", style: .default, handler: { _ in
                         localDB().writeSettings(item: "isOffline", value: "1")
                         self.type = "download"
@@ -157,24 +155,18 @@ class OfflineSettingsViewController: UITableViewController, VersionProtocol, loc
             
         }else{
             let alert = UIAlertController(title: nil, message: nil, preferredStyle: .actionSheet)
-            alert.addAction(UIAlertAction(title: "刪除離線資料檔（包括已下載圖片）", style: .destructive, handler: { _ in
+            alert.addAction(UIAlertAction(title: "刪除離線資料檔", style: .destructive, handler: { _ in
                 localDB().writeSettings(item: "isOffline", value: "0")
                 localDB().delete(self)
                 self.removeImg()
                 self.localVersion.text = " "
                 self.localTS.text = " "
-                self.userDefaults.set(false, forKey: "offlineImg")
-                self.downImgSwitch.setOn(false, animated: true)
                 self.localCheck.isEnabled = false
-                self.downImgSwitch.isEnabled = false
             }))
-            alert.addAction(UIAlertAction(title: "保留離線資料檔（將刪除已下載圖片）", style: .default, handler: { _ in
+            alert.addAction(UIAlertAction(title: "保留離線資料檔", style: .default, handler: { _ in
                 localDB().writeSettings(item: "isOffline", value: "0")
-                self.userDefaults.set(false, forKey: "offlineImg")
-                self.downImgSwitch.setOn(false, animated: true)
                 self.removeImg()
                 self.localCheck.isEnabled = false
-                self.downImgSwitch.isEnabled = false
             }))
             alert.addAction(UIAlertAction(title: "取消", style: .cancel, handler: { _ in
                 self.switchView.setOn(true, animated: true)
@@ -182,18 +174,21 @@ class OfflineSettingsViewController: UITableViewController, VersionProtocol, loc
             self.present(alert, animated: true, completion: nil)
         }
     }
-    func donwloadImg(){
-        imgAlert.dismiss(animated: true, completion: nil)
-        prog.progress = 0.0
-        prog.progressViewStyle = .bar
-        prog.tintColor = UIColor(red: 0.0, green: 122/255, blue: 255/255, alpha: 1.0)
-        startAlert.view.addSubview(prog)
-        self.present(startAlert, animated: true, completion: nil)
-        let db = Session.sharedInstance.db
-        if let mydb = db{
-            let statement = mydb.fetch("info", cond: nil, order: nil)
-            while sqlite3_step(statement) == SQLITE_ROW{
-                let cover = String(cString: sqlite3_column_text(statement, 0)) + ".jpg"
+    func parseImg(_ data: Data, mode: String){
+        var jsonResult = NSArray()
+        var jsonElement = NSDictionary()
+        do{
+            jsonResult = try JSONSerialization.jsonObject(with: data, options:JSONSerialization.ReadingOptions.allowFragments) as! NSArray
+        } catch let error as NSError {
+            print(error)
+        }
+        for i in 0 ..< jsonResult.count
+        {
+            jsonElement = jsonResult[i] as! NSDictionary
+            
+            guard let id = jsonElement["ID"] else {return}
+            let cover = id as! String + ".jpg"
+            if mode == "download"{
                 let url = URL(string: "https://scarletsc.net/girlfrontline/img/" + cover)
                 let filePath = self.localPath[self.localPath.count-1].absoluteString + cover
                 DownloadPhoto().get(url: url!) { data, response, error in
@@ -208,8 +203,62 @@ class OfflineSettingsViewController: UITableViewController, VersionProtocol, loc
                         }
                     })
                 }
+            }else if mode == "delete"{
+                let documentsPath = NSSearchPathForDirectoriesInDomains(.documentDirectory,.userDomainMask,true)[0] as NSString
+                let destinationPath = documentsPath.appendingPathComponent(cover)
+                do{
+                    try FileManager.default.removeItem(atPath: destinationPath)
+                }catch{
+                    print(error)
+                }
             }
+            
         }
+    }
+    func donwloadImg(){
+        imgAlert.dismiss(animated: true, completion: nil)
+        prog.progress = 0.0
+        prog.progressViewStyle = .bar
+        prog.tintColor = UIColor(red: 0.0, green: 122/255, blue: 255/255, alpha: 1.0)
+        startAlert.view.addSubview(prog)
+        self.present(startAlert, animated: true, completion: nil)
+        if localDB().readSettings()[0]{
+            let db = Session.sharedInstance.db
+            if let mydb = db{
+                let statement = mydb.fetch("info", cond: nil, order: nil)
+                while sqlite3_step(statement) == SQLITE_ROW{
+                    let cover = String(cString: sqlite3_column_text(statement, 0)) + ".jpg"
+                    let url = URL(string: "https://scarletsc.net/girlfrontline/img/" + cover)
+                    let filePath = self.localPath[self.localPath.count-1].absoluteString + cover
+                    DownloadPhoto().get(url: url!) { data, response, error in
+                        guard let imgData = data, error == nil else { return }
+                        DispatchQueue.main.async(execute: { () -> Void in
+                            let img = UIImage(data: imgData)?.jpegData(compressionQuality: 1.0)
+                            do{
+                                try img?.write(to: URL(string: filePath)!)
+                                self.prog.progress += 0.04
+                            }catch{
+                                print(error)
+                            }
+                        })
+                    }
+                }
+            }
+        }else{
+            let url: URL = URL(string: "https://scarletsc.net/girlfrontline/search.php")!
+            let defaultSession = URLSession(configuration: URLSessionConfiguration.default, delegate: self, delegateQueue: nil )
+            let task = defaultSession.dataTask(with: url) {
+                (data, response, error) in
+                if error != nil {
+                    print("Failed to download data")
+                }else {
+                    print("Data downloaded")
+                    self.parseImg(data!, mode: "download")
+                }
+            }
+            task.resume()
+        }
+        
         startAlert.dismiss(animated: true, completion: {
             let alert = UIAlertController(title: "下載完成", message: nil, preferredStyle: .alert)
             alert.addAction(UIAlertAction(title: "確定", style: .default, handler: nil))
@@ -220,22 +269,37 @@ class OfflineSettingsViewController: UITableViewController, VersionProtocol, loc
     }
     func removeImg(){
         delImgAlert.dismiss(animated: true, completion: nil)
-        let db = Session.sharedInstance.db
-        if let mydb = db{
-            let statement = mydb.fetch("info", cond: nil, order: nil)
-            while sqlite3_step(statement) == SQLITE_ROW{
-                let cover = String(cString: sqlite3_column_text(statement, 0)) + ".jpg"
-                let documentsPath = NSSearchPathForDirectoriesInDomains(.documentDirectory,.userDomainMask,true)[0] as NSString
-                let destinationPath = documentsPath.appendingPathComponent(cover)
-//                let filePath = self.localPath[self.localPath.count-1].absoluteString + cover
-                print(destinationPath)
-                do{
-                    try FileManager.default.removeItem(atPath: destinationPath)
-                }catch{
-                    print(error)
+        
+        if localDB().readSettings()[0]{
+            let db = Session.sharedInstance.db
+            if let mydb = db{
+                let statement = mydb.fetch("info", cond: nil, order: nil)
+                while sqlite3_step(statement) == SQLITE_ROW{
+                    let cover = String(cString: sqlite3_column_text(statement, 0)) + ".jpg"
+                    let documentsPath = NSSearchPathForDirectoriesInDomains(.documentDirectory,.userDomainMask,true)[0] as NSString
+                    let destinationPath = documentsPath.appendingPathComponent(cover)
+                    do{
+                        try FileManager.default.removeItem(atPath: destinationPath)
+                    }catch{
+                        print(error)
+                    }
                 }
             }
+        }else{
+            let url: URL = URL(string: "https://scarletsc.net/girlfrontline/search.php")!
+            let defaultSession = URLSession(configuration: URLSessionConfiguration.default, delegate: self, delegateQueue: nil )
+            let task = defaultSession.dataTask(with: url) {
+                (data, response, error) in
+                if error != nil {
+                    print("Failed to download data")
+                }else {
+                    print("Data downloaded")
+                    self.parseImg(data!, mode: "delete")
+                }
+            }
+            task.resume()
         }
+        
         self.userDefaults.set(false, forKey: "offlineImg")
         let alert = UIAlertController(title: "已刪除", message: nil, preferredStyle: .alert)
         alert.addAction(UIAlertAction(title: "確定", style: .default, handler: nil))
@@ -272,11 +336,6 @@ class OfflineSettingsViewController: UITableViewController, VersionProtocol, loc
             self.downImgSwitch.setOn(true, animated: true)
         }))
         
-        if userDefaults.object(forKey: "offlineImg") != nil{
-        }else{
-            userDefaults.set(false, forKey: "offlineImg")
-        }
-        
         switchView.setOn(localDB().readSettings()[0], animated: true)
         switchView.tag = 1 // for detect which row switch Changed
         switchView.addTarget(self, action: #selector(self.switchChanged(_:)), for: .valueChanged)
@@ -297,12 +356,9 @@ class OfflineSettingsViewController: UITableViewController, VersionProtocol, loc
         if localDB().readSettings()[0]{
             ver.localVersion()
             localCheck.isEnabled = true
-            downImgSwitch.isEnabled = true
         }else{
             localCheck.isEnabled = false
-            downImgSwitch.isEnabled = false
         }
-        
         let db = Session.sharedInstance.db
         if let mydb = db{
             let statement = mydb.fetch("info", cond: nil, order: nil)
@@ -310,7 +366,6 @@ class OfflineSettingsViewController: UITableViewController, VersionProtocol, loc
                 ver.localVersion()
             }
         }
-        
     }
 
     // MARK: - Table view data source
@@ -320,11 +375,9 @@ class OfflineSettingsViewController: UITableViewController, VersionProtocol, loc
     override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
         switch section {
         case 0:
-            return 1
-        case 1:
+            return 2
+        case 1, 2:
             return 3
-        case 2:
-            return 4
         default:
             return 0
         }
